Revert "cleanup: move source code component from debugger-v2 to widgets (#4396)"
This reverts commit db71ff05181a80038711f9be74572f9feabfa84f.
