+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/source_code/source_code_container.ts
@@ -36,7 +36,7 @@ const windowWithRequireAndMonaco: any = window;
  * time.
  */
 @Component({
-  selector: 'source-code',
+  selector: 'tf-debugger-v2-source-code',
   template: `
     <source-code-component
       [lines]="lines"
