+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/source_files/source_files_component.css
@@ -55,7 +55,7 @@ limitations under the License.
   vertical-align: top;
 }
 
-source-code {
+tf-debugger-v2-source-code {
   flex-grow: 1;
   width: 100%;
 }
