+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/source_files/source_files_component.ng.html
@@ -35,12 +35,12 @@ limitations under the License.
     </ng-template>
   </div>
 
-  <source-code
+  <tf-debugger-v2-source-code
     *ngIf="focusedSourceFileContent !== null &&
            focusedSourceLineSpec !== null &&
            focusedSourceFileContent.lines !== null"
     [lines]="focusedSourceFileContent.lines"
     [focusedLineno]="focusedSourceLineSpec.lineno"
   >
-  </source-code>
+  </tf-debugger-v2-source-code>
 </div>
