+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/stack_trace/BUILD
@@ -19,7 +19,7 @@ tf_ng_module(
         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/actions",
         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store",
         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store:types",
-        "//tensorboard/webapp/widgets/source_code",
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/source_code",
         "@npm//@angular/common",
         "@npm//@angular/core",
         "@npm//@ngrx/store",
