+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/stack_trace/stack_trace_module.ts
@@ -16,7 +16,7 @@ limitations under the License.
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 
-import {SourceCodeModule} from '../../../../../webapp/widgets/source_code/source_code_module';
+import {SourceCodeModule} from '../source_code/source_code_module';
 import {StackTraceComponent} from './stack_trace_component';
 import {StackTraceContainer} from './stack_trace_container';
